backend/pyproject.toml
[project]
name = "stonks-backend"
version = "0.1.0"
description = "Stonks autonomous options paper trading backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.14.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "httpx>=0.27.0",
  "websockets>=13.0",
  "python-dotenv>=1.0.1",
  "apscheduler>=3.10.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
